src/requirements.txt
Django==3.0.5
psycopg2==2.8.5
django-crispy-forms==1.9.0
requests==2.23.0
django-service-objects==0.6.0
djangorestframework==3.11.0
